Add a --watch flag for event-driven formatting.
This adds a `format --watch .` mode which uses inotify to run the
formatter on modifications to any of the watched files. Currently only
files that exist at the time that the program begins are watched. In
the future it would be nice to recursively watch directory arguments
and add watchers for new files.

github.com/ChrisCummins/format/issues/12

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>

[Exported from ea7bf18a0f65d22e88f3ac59ece6d4ccda24abaf]